CryptoLibrary/requirements.txt
asgiref==3.7.2
cffi==1.15.1
colorama==0.4.6
contourpy==1.2.0
crayons==0.4.0
cryptography==41.0.2
cycler==0.12.1
Django==4.2.4
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.47.0
imageio==2.33.1
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.7.1
networkx==3.2.1
numpy==1.25.2
opencv-python==4.8.0.74
packaging==23.2
piexif==1.1.3
Pillow==9.5.0
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-magic-bin==0.4.14
pytz==2023.3
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
stegano==0.11.2
tifffile==2023.12.9
tzdata==2023.3
